requirements-dev.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cv2-tools==2.4.0
cycler==0.12.1
Django==4.2.9
flatbuffers==23.5.26
fonttools==4.49.0
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
ImageHash==4.3.1
install==1.3.5
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-constraint==1.4.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyWavelets==1.5.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
